+++ src/main/java/com/codenear/butterfly/payment/domain/repository/CancelPaymentRepository.java
@@ -1,4 +1,4 @@
-package com.codenear.butterfly.payment.kakaoPay.domain.repository;
+package com.codenear.butterfly.payment.domain.repository;
 
 import com.codenear.butterfly.payment.domain.CancelPayment;
 import org.springframework.data.jpa.repository.JpaRepository;
