+++ src/main/java/com/codenear/butterfly/payment/kakaoPay/application/KakaoPaymentServiceImpl.java
@@ -10,12 +10,12 @@ import com.codenear.butterfly.payment.domain.dto.OrderType;
 import com.codenear.butterfly.payment.domain.dto.PaymentStatus;
 import com.codenear.butterfly.payment.domain.dto.handler.ApprovePaymentHandler;
 import com.codenear.butterfly.payment.domain.dto.request.BasePaymentRequestDTO;
+import com.codenear.butterfly.payment.domain.repository.CancelPaymentRepository;
 import com.codenear.butterfly.payment.domain.repository.OrderDetailsRepository;
 import com.codenear.butterfly.payment.domain.repository.PaymentRedisRepository;
+import com.codenear.butterfly.payment.domain.repository.SinglePaymentRepository;
 import com.codenear.butterfly.payment.kakaoPay.domain.dto.ApproveResponseDTO;
 import com.codenear.butterfly.payment.kakaoPay.domain.dto.ReadyResponseDTO;
-import com.codenear.butterfly.payment.kakaoPay.domain.repository.CancelPaymentRepository;
-import com.codenear.butterfly.payment.kakaoPay.domain.repository.SinglePaymentRepository;
 import com.codenear.butterfly.payment.kakaoPay.util.KakaoPaymentUtil;
 import com.codenear.butterfly.point.domain.PointRepository;
 import com.codenear.butterfly.product.domain.ProductInventory;
@@ -29,22 +29,22 @@ import java.util.UUID;
 
 @Transactional(readOnly = true)
 @Service
-public class SinglePaymentService extends PaymentService implements KakaoPaymentService {
+public class KakaoPaymentServiceImpl extends PaymentService implements KakaoPaymentService {
     private final PaymentRedisRepository kakaoPaymentRedisRepository;
     private final KakaoPaymentUtil<Object> kakaoPaymentUtil;
     private final ProductInventoryRepository productInventoryRepository;
 
-    public SinglePaymentService(SinglePaymentRepository singlePaymentRepository,
-                                AddressRepository addressRepository,
-                                OrderDetailsRepository orderDetailsRepository,
-                                MemberRepository memberRepository,
-                                ProductInventoryRepository productInventoryRepository,
-                                PaymentRedisRepository kakaoPaymentRedisRepository,
-                                PointRepository pointRepository,
-                                ApplicationEventPublisher applicationEventPublisher,
-                                KakaoPaymentUtil<Object> kakaoPaymentUtil,
-                                FCMFacade fcmFacade,
-                                CancelPaymentRepository cancelPaymentRepository) {
+    public KakaoPaymentServiceImpl(SinglePaymentRepository singlePaymentRepository,
+                                   AddressRepository addressRepository,
+                                   OrderDetailsRepository orderDetailsRepository,
+                                   MemberRepository memberRepository,
+                                   ProductInventoryRepository productInventoryRepository,
+                                   PaymentRedisRepository kakaoPaymentRedisRepository,
+                                   PointRepository pointRepository,
+                                   ApplicationEventPublisher applicationEventPublisher,
+                                   KakaoPaymentUtil<Object> kakaoPaymentUtil,
+                                   FCMFacade fcmFacade,
+                                   CancelPaymentRepository cancelPaymentRepository) {
         super(singlePaymentRepository, addressRepository, orderDetailsRepository, memberRepository, productInventoryRepository, kakaoPaymentRedisRepository, pointRepository, applicationEventPublisher, fcmFacade, cancelPaymentRepository);
         this.kakaoPaymentUtil = kakaoPaymentUtil;
         this.kakaoPaymentRedisRepository = kakaoPaymentRedisRepository;
