+++ src/main/java/com/codenear/butterfly/payment/tossPay/application/TossPaymentServiceImpl.java
@@ -17,11 +17,11 @@ import com.codenear.butterfly.payment.domain.dto.handler.CancelHandler;
 import com.codenear.butterfly.payment.domain.dto.handler.CancelPaymentHandler;
 import com.codenear.butterfly.payment.domain.dto.request.BasePaymentRequestDTO;
 import com.codenear.butterfly.payment.domain.dto.request.CancelRequestDTO;
+import com.codenear.butterfly.payment.domain.repository.CancelPaymentRepository;
 import com.codenear.butterfly.payment.domain.repository.OrderDetailsRepository;
 import com.codenear.butterfly.payment.domain.repository.PaymentRedisRepository;
+import com.codenear.butterfly.payment.domain.repository.SinglePaymentRepository;
 import com.codenear.butterfly.payment.exception.PaymentException;
-import com.codenear.butterfly.payment.kakaoPay.domain.repository.CancelPaymentRepository;
-import com.codenear.butterfly.payment.kakaoPay.domain.repository.SinglePaymentRepository;
 import com.codenear.butterfly.payment.tossPay.domain.dto.CancelResponseDTO;
 import com.codenear.butterfly.payment.tossPay.domain.dto.ConfirmResponseDTO;
 import com.codenear.butterfly.payment.tossPay.domain.dto.ReadyResponseDTO;
